BugFixes and Tweaks
Lots of fixes and upgrades to Config. Config.sync() now properly handles
syncing ALL data bewteen the various processes. There are some
inefficiencies with transferring strings but we live with them as the
main slowdown in the code will never be MPI communication...so we hope.

Added deprecated directory to git tracking. This is probably bad
practice but I want to keep these files around without having to worry
about compiling them and satisfying dependencies.