Removed vector.clear()/vector.reserve()/vector.push_back() combo
Now we just do vector.resize() and then directly set the member of the
vector via index. This seems like a slightly less c++ way to do things,
but since resize is an effective no-op if the vector is already the
correct size, this should save some cycles during the frame loop.