Created testing directory
This is an half-assed attempt at having a proper testing suite. For now
we'll just a dir where we keep  various mini-testing environments. In
this case, we have a main file that allows us to open and call various
AtomGroup commands so that we can test the AtomGroup classes along with
the Reader classes.