Kernel Switcher Implemented
Along with other minor aesthetic changes to cmdline output, the
kernel_switcher function has been implemented. This makes the stepper
function a bit smaller and reduces the clutter in that function. It
further paves the way for the Writer class as well which will further
clean up this function call.