Timing in stepper was Wonky
Added a Barrier() at the end of the frame loop in order to make the
atomgroup timings more normal. Bascially, the asynchronicities in the
timing was showing up in the atomgroup column which was causing us to
think that I/O was an performance issue rather than it just being that
the last MPI process had a larger load and was taking significantly
longer to finish running the calculation kernel.